# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        },
        "summary": "Delete user by ID",
        "tags": [
          "User"
        ]
      }
    },
...
    "/agent/{id}": {
      "get": {
        "operationId": "AgentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details"
          }
        },
        "summary": "Get agent by ID",
        "tags": [
          "Agent"
        ]
      },
      "patch": {
        "operationId": "AgentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent updated"
          }
        },
        "summary": "Update agent details",
        "tags": [
          "Agent"
        ]
      },
      "delete": {
        "operationId": "AgentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent removed"
          }
        },
        "summary": "Delete agent (soft delete)",
        "tags": [
          "Agent"
        ]
      }
    },
...
    "/admin/projects/custom/{submissionId}/approve": {
      "patch": {
        "operationId": "AdminController_approveCustomProject",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "description": "Submission UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Custom project approved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Approve custom design project submission",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/projects/custom/{submissionId}/reject": {
      "patch": {
        "operationId": "AdminController_rejectCustomProject",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "description": "Submission UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProjectActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Custom project rejected"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Reject custom design project submission",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/custom-design/initialize": {
      "post": {
        "operationId": "CustomDesignController_initialize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeCustomDesignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Custom design session initialized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Step 1: Initialize a new custom design wizard session",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design": {
      "post": {
        "description": "Uploads all wizard steps and optional reference files at once",
        "operationId": "CustomDesignController_createAndSubmit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCustomDesignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Custom design project submitted"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "One-shot submission: Complete wizard and submit custom design project",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/my": {
      "get": {
        "operationId": "CustomDesignController_getMyDesigns",
        "parameters": [
          {
            "name": "serviceType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "landscape-design",
                "interior-design",
                "product-design",
                "3d-rendering",
                "2d-architectural",
                "building-information-modeling",
                "mep-drawings",
                "structural-engineering"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "in_progress",
                "submitted",
                "under_review",
                "approved",
                "rejected",
                "completed",
                "revision"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user custom designs"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user custom design projects",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/config/{serviceType}": {
      "get": {
        "description": "Returns available options, building contexts, scopes, and validation rules for the given service type",
        "operationId": "CustomDesignController_getServiceConfig",
        "parameters": [
          {
            "name": "serviceType",
            "required": true,
            "in": "path",
            "description": "Service type enum e.g. Architectural, Structural, MEP, Interior",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service configuration returned"
          }
        },
        "summary": "Get dynamic wizard configuration for a service type (Public)",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/user/{userId}": {
      "get": {
        "operationId": "CustomDesignController_getByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "landscape-design",
                "interior-design",
                "product-design",
                "3d-rendering",
                "2d-architectural",
                "building-information-modeling",
                "mep-drawings",
                "structural-engineering"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "in_progress",
                "submitted",
                "under_review",
                "approved",
                "rejected",
                "completed",
                "revision"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Custom design list"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get custom designs by user ID (Admin/Staff)",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/agent/{agentId}": {
      "get": {
        "operationId": "CustomDesignController_getByAgent",
        "parameters": [
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "description": "Agent ID UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "landscape-design",
                "interior-design",
                "product-design",
                "3d-rendering",
                "2d-architectural",
                "building-information-modeling",
                "mep-drawings",
                "structural-engineering"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "in_progress",
                "submitted",
                "under_review",
                "approved",
                "rejected",
                "completed",
                "revision"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Custom design list"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get custom designs assigned to an agent",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/{id}": {
      "get": {
        "operationId": "CustomDesignController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom Design ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Custom design project details"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get custom design project details by ID",
        "tags": [
          "Custom Design"
        ]
      },
      "patch": {
        "operationId": "CustomDesignController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom Design ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomDesignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Custom design updated"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update custom design draft",
        "tags": [
          "Custom Design"
        ]
      },
      "delete": {
        "operationId": "CustomDesignController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom Design ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Custom design deleted"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete custom design project",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/{id}/step": {
      "patch": {
        "operationId": "CustomDesignController_saveStep",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom Design ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SaveStepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Step data saved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Save progress for a specific wizard step",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/{id}/submit": {
      "post": {
        "operationId": "CustomDesignController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom Design ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draft submitted successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Finalize and submit a saved draft project",
        "tags": [
          "Custom Design"
        ]
      }
    },
    "/custom-design/{id}/validate-step": {
      "post": {
        "operationId": "CustomDesignController_validateStep",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom Design ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Step validation status"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Validate single step payload against service rules",
        "tags": [
          "Custom Design"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mejarc API",
    "description": "Comprehensive REST API documentation for Mejarc Backend services, including Authentication, User Management, Agent Management, Marketplace, Custom Design Wizard, Orders, Payments, Wallet & Withdrawals, Notifications, and Real-Time Chat.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your Bearer token in format: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "AdminProjectActionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "InitializeCustomDesignDto": {
        "type": "object",
        "properties": {
          "serviceType": {
            "type": "string",
            "enum": [
              "landscape-design",
              "interior-design",
              "product-design",
              "3d-rendering",
              "2d-architectural",
              "building-information-modeling",
              "mep-drawings",
              "structural-engineering"
            ]
          },
          "serviceContext": {
            "type": "string"
          },
          "selectionMethod": {
            "type": "string",
            "enum": [
              "manual",
              "template",
              "ai-advice",
              "contest",
              "hire_pro",
              "fixed_quote"
            ]
          }
        },
        "required": [
          "serviceType",
          "serviceContext"
        ]
      },
...
    }
  }
}